finished basic login
