+++ client/src/App.jsx
@@ -10,7 +10,6 @@ import 'bootstrap/dist/css/bootstrap.min.css'
 
 import { StoreProvider } from './utils/store-context';
 import Nav from './components/Nav';
-import Login from './components/Login';
 import Auth from './utils/auth'
 
 import './app.scss';
@@ -39,7 +38,7 @@ function App() {
       <StoreProvider>
         <div id="app-shell">
           <Nav />
-          {Auth.loggedIn() ? <Outlet /> : <Login />}
+          <Outlet /> 
         </div>
       </StoreProvider>
     </ApolloProvider>
